StorageProvider Fix: skip if doesn't exit in copyWithoutConfirm

+++ src/java/org/apache/cassandra/io/util/FileUtils.java
@@ -242,7 +242,8 @@ public final class FileUtils
     {
         try
         {
-            Files.copy(from.toPath(), to.toPath());
+            if (from.exists())
+                Files.copy(from.toPath(), to.toPath());
         }
         catch (IOException e)
         {
